requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
pandas>=2.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
streamlit>=1.39.0
anthropic>=0.39.0
openai>=1.52.0
google-generativeai>=0.8.0
